+++ feature/input/src/main/java/com/freshdigitable/udonroad2/input/TweetInputFragment.kt
@@ -16,7 +16,6 @@
 
 package com.freshdigitable.udonroad2.input
 
-import android.animation.ValueAnimator
 import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -26,15 +25,9 @@ import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
 import android.view.inputmethod.InputMethodManager
-import androidx.core.animation.doOnEnd
-import androidx.core.animation.doOnStart
-import androidx.core.view.doOnPreDraw
-import androidx.core.view.updateLayoutParams
-import androidx.databinding.BindingAdapter
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import androidx.interpolator.view.animation.FastOutSlowInInterpolator
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.navArgs
 import com.freshdigitable.udonroad2.input.databinding.FragmentTweetInputBinding
@@ -81,12 +74,10 @@ class TweetInputFragment : Fragment() {
             requireActivity().invalidateOptionsMenu()
         }
         binding.twIntext.setOnFocusChangeListener { v, hasFocus ->
-            val inputMethod =
-                v.context.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
             if (!hasFocus && viewModel.isExpanded.value == false) {
-                inputMethod.hideSoftInputFromWindow(v.windowToken, 0)
+                v.hideInputMethod()
             } else {
-                inputMethod.showSoftInput(v, InputMethodManager.SHOW_FORCED)
+                v.showInputMethod()
             }
         }
         viewLifecycleOwner.lifecycleScope.launchWhenResumed {
@@ -126,92 +117,15 @@ class TweetInputFragment : Fragment() {
     }
 }
 
-enum class InputMenuItem(
-    val itemId: Int,
-    val enabled: Boolean
-) {
-    WRITE_ENABLED(R.id.input_tweet_write, true),
-    WRITE_DISABLED(R.id.input_tweet_write, false),
-    SEND_ENABLED(R.id.input_tweet_send, true),
-    SEND_DISABLED(R.id.input_tweet_send, false),
-    RETRY_ENABLED(R.id.input_tweet_error, true),
-}
-
-fun Menu.prepareItem(available: InputMenuItem) {
-    InputMenuItem.values().map { it.itemId }.distinct().forEach {
-        val item = findItem(it)
-        when (item.itemId) {
-            available.itemId -> {
-                item.isVisible = true
-                item.isEnabled = available.enabled
-            }
-            else -> {
-                item.isVisible = false
-            }
-        }
+private val Context.inputMethodManager: InputMethodManager
+    get() {
+        return getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
     }
+
+private fun View.showInputMethod() {
+    context.inputMethodManager.showSoftInput(this, InputMethodManager.SHOW_FORCED)
 }
 
-@BindingAdapter("isExpanded", "onExpandAnimationEnd", requireAll = false)
-fun View.expand(isExpanded: Boolean?, onExpandAnimationEnd: (() -> Unit)?) {
-    when (isExpanded) {
-        true -> setupExpendAnim(onExpandAnimationEnd)
-        else -> collapseWithAnim()
-    }
-}
-
-private fun View.setupExpendAnim(onExpandAnimEnd: (() -> Unit)?) {
-    measure(
-        View.MeasureSpec.makeMeasureSpec(0, View.MeasureSpec.UNSPECIFIED),
-        View.MeasureSpec.makeMeasureSpec(0, View.MeasureSpec.UNSPECIFIED)
-    )
-    when {
-        measuredHeight > 0 -> expandWithAnim(onExpandAnimEnd)
-        else -> doOnPreDraw { it.expandWithAnim(onExpandAnimEnd) }
-    }
-}
-
-private fun View.expandWithAnim(onExpandAnimEnd: (() -> Unit)?) {
-    val container = parent as View
-    val h = measuredHeight
-    ValueAnimator.ofInt(-h, 0).apply {
-        duration = 200
-        interpolator = FastOutSlowInInterpolator()
-        doOnStart {
-            this@expandWithAnim.visibility = View.VISIBLE
-        }
-        addUpdateListener {
-            val animValue = it.animatedValue as Int
-            this@expandWithAnim.translationY = animValue.toFloat()
-            container.updateLayoutParams {
-                height = h + animValue
-            }
-        }
-        doOnEnd {
-            this@expandWithAnim.translationY = 0f
-            container.updateLayoutParams {
-                height = ViewGroup.LayoutParams.WRAP_CONTENT
-            }
-            onExpandAnimEnd?.invoke()
-        }
-    }.start()
-}
-
-fun View.collapseWithAnim() {
-    val container = parent as View
-    val h = measuredHeight
-    ValueAnimator.ofInt(-h).apply {
-        duration = 200
-        interpolator = FastOutSlowInInterpolator()
-        addUpdateListener {
-            val animValue = it.animatedValue as Int
-            this@collapseWithAnim.translationY = animValue.toFloat()
-            container.updateLayoutParams {
-                height = h + animValue
-            }
-        }
-        doOnEnd {
-            this@collapseWithAnim.visibility = View.GONE
-        }
-    }.start()
+private fun View.hideInputMethod() {
+    context.inputMethodManager.hideSoftInputFromWindow(windowToken, 0)
 }
